third_party/blink/web_tests/external/wpt/speculation-rules/prefetch/single-url.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>
<script>
promise_test(async t => {
  assert_implements(HTMLScriptElement.supports('speculationrules'), "Speculation Rules not supported");

  let agent = await spawnWindow(t);
  let nextUrl = agent.getExecutorURL({page: 2});
  await agent.forceSinglePrefetch(nextUrl);
  await agent.navigate(nextUrl);

  let requestHeaders = await agent.getRequestHeaders();
  assert_in_array(requestHeaders.purpose, ["", "prefetch"], "The vendor-specific header Purpose, if present, must be 'prefetch'.");
  assert_equals(requestHeaders.sec_purpose, "prefetch");
}, "test single url prefetch");
</script>
